end of 1.03 assignment, added player generation, and distance values from player for every location in the map

+++ 327rlg/headers/gameio.h
@@ -24,7 +24,7 @@
  * @author Collin Vincent <collinvincent96@gmail.com>
  * @date   2017-02-01T20:23:07+000
  */
-int saveMap(int numRooms, room_t const *rooms, uint8_t map_hard[mapHeight][mapWidth]);
+int saveMap(uint8_t numRooms, room_t const *rooms, uint8_t map_hard[mapHeight][mapWidth]);
 
 /**
  * loads a map from the files
@@ -34,7 +34,7 @@ int saveMap(int numRooms, room_t const *rooms, uint8_t map_hard[mapHeight][mapWi
  * @author Collin Vincent <collinvincent96@gmail.com>
  * @date   2017-02-01T20:26:11+000
  */
-int loadMap(room_t **rooms, int *numRooms, uint8_t map_hard[mapHeight][mapWidth]);
+int loadMap(room_t **rooms, uint8_t *numRooms, uint8_t map_hard[mapHeight][mapWidth]);
 
 
 #endif
